+++ config/config.ini
@@ -1,8 +1,9 @@
 [GUI]
 column_width = 100
-two_window_mode = True
+two_window_mode = False
 resolution_width = 1920
 resolution_height = 1080
+dark_mode = True
 
 [Import]
 separator = ,
